# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Note App API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Авторизация пользователя(получение токена)",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "email": {
                                                "type": "string",
                                                "example": "ivanov@mail.ru"
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": "12345678"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Ivan Ivanov"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "ivanov@mail.ru"
                                                },
                                                "email_verified_at": {
                                                    "type": "timestamp",
                                                    "example": null
                                                },
                                                "created_at": {
                                                    "type": "timestamp",
                                                    "example": "2023-11-28T18:56:11.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "timestamp",
                                                    "example": "2023-11-28T18:56:11.000000Z"
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "example": "user"
                                                },
                                                "token": {
                                                    "type": "string",
                                                    "example": "tWSGolP01gKzmIQVFP7fHe7s970wv7H6vvvYeeaDOE6dOmCXoqdzSImdV9s-_wJb4IrslyeNbxU4SFeiEhMS1TDpBfjjA6y1H7IuOGAvUgry1zmv1240gy--vmdikomjFKrXnylTKhbKBVFeZG5wbNkpGBrcLAtCTRpZ2oqmDOsLkvpp0Ed_WrClbFV9O4NuoQ"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Регистрация пользователя",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Ivan Ivanov"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "ivanov@mail.ru"
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": "12345678"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User created successfully."
                                        },
                                        "user": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "Ivan Ivanov"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "ivanov@mail.ru"
                                                },
                                                "updated_at": {
                                                    "type": "timestamp",
                                                    "example": "2023-11-28T18:56:11.000000Z"
                                                },
                                                "created_at": {
                                                    "type": "timestamp",
                                                    "example": "2023-11-28T18:56:11.000000Z"
                                                },
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Выйти из системы",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully logged out"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}